refactor: OrderTemplate에서, 하단부를 컴포넌트로 분리
분리 전에 OrderTemplate의 일이 너무 많았음
